+++ tests/neighbourhood.lua
@@ -16,8 +16,8 @@ function testNeighbourhood:testVonNeumann()
      lu.assertEquals(4, #von_neumann )
      lu.assertEquals(16, #von_neumann.categories )
      -- Test categorisation
-     local ct1 = von_neumann:categorise(self.pattern_1, self.pattern_1:com())
-     local ct2 = von_neumann:categorise(self.pattern_2, self.pattern_2:com())
+     local ct1 = von_neumann:categorise(self.pattern_1, self.pattern_1:medoid())
+     local ct2 = von_neumann:categorise(self.pattern_2, self.pattern_2:medoid())
      lu.assertEquals(16,ct1) -- Lowest category (single cell)
      lu.assertEquals(1, ct2) -- Highest category (full neighbourhood)
 end
@@ -27,8 +27,8 @@ function testNeighbourhood:testMoore()
      lu.assertEquals(8, #moore)
      lu.assertEquals(256, #moore.categories)
      -- Test categorisation
-     local ct1 = moore:categorise(self.pattern_1, self.pattern_1:com())
-     local ct2 = moore:categorise(self.pattern_2, self.pattern_2:com())
+     local ct1 = moore:categorise(self.pattern_1, self.pattern_1:medoid())
+     local ct2 = moore:categorise(self.pattern_2, self.pattern_2:medoid())
      lu.assertEquals(256,ct1) -- Lowest category (single cell)
      lu.assertEquals(1, ct2)  -- Highest category (full neighbourhood)
 end