+++ tests/pattern.lua
@@ -50,19 +50,83 @@ function testPattern:testConstructor()
     lu.assertEquals(self.pattern_5, self.pattern_2)
 end
 
+function testPattern:testClone()
+    -- Test pattern cloning
+    local pattern_5_clone = self.pattern_5:clone()
+    lu.assertEquals(pattern_5_clone, self.pattern_5)
+    lu.assertNotEquals(pattern_5_clone, self.pattern_4)
+end
+
 -- Test insert methods
--- Note insert *mutates* patterns, so here we test a clone
 function testPattern:testInsert()
     -- Test both insert methods
-    local pattern_1_clone = self.pattern_1:clone()
-    pattern.insert(pattern_1_clone, 1, -1)
-    pattern_1_clone:insert(-1, 1)
+    local insert_test = pattern.new()
+    pattern.insert(insert_test, 1, -1)
+    insert_test:insert(-1, 1)
 
-    lu.assertEquals(pattern.size(pattern_1_clone),2)
-    lu.assertEquals(pattern_1_clone.max.x,1)
-    lu.assertEquals(pattern_1_clone.max.y,1)
-    lu.assertEquals(pattern_1_clone.min.x,-1)
-    lu.assertEquals(pattern_1_clone.min.y,-1)
+    lu.assertEquals(pattern.size(insert_test),2)
+    lu.assertEquals(insert_test.max.x,1)
+    lu.assertEquals(insert_test.max.y,1)
+    lu.assertEquals(insert_test.min.x,-1)
+    lu.assertEquals(insert_test.min.y,-1)
+end
+
+function testPattern:testCentroid()
+    -- Test five patterns which should have the same
+    -- centroid, and one which should not
+    local centroid1 = pattern.new({{1,1,1},
+                                   {1,0,1},
+                                   {1,1,1}}):centroid()
+    local centroid2 = pattern.new({{1,0,1},
+                                   {0,0,0},
+                                   {1,0,1}}):centroid()
+    local centroid3 = pattern.new({{0,1,0},
+                                   {0,0,0},
+                                   {0,1,0}}):centroid()
+    local centroid4 = pattern.new({{0,0,0},
+                                   {1,0,1},
+                                   {0,0,0}}):centroid()
+    local centroid5 = pattern.new({{0,0,0},
+                                   {0,1,0},
+                                   {0,0,0}}):centroid()
+    -- This should not be the same as the others
+    local centroid6 = pattern.new({{1,1,1},
+                                   {0,0,0},
+                                   {0,0,0}}):centroid()
+    lu.assertEquals(centroid1, centroid2)
+    lu.assertEquals(centroid2, centroid3)
+    lu.assertEquals(centroid3, centroid4)
+    lu.assertEquals(centroid4, centroid5)
+    lu.assertNotEquals(centroid5, centroid6)
+end
+
+function testPattern:testMedoid()
+    -- Test five patterns which should have the same
+    -- medoid, and one which should not
+    local medoid1 = pattern.new({{1,1,1},
+                                 {1,1,1},
+                                 {1,1,1}}):medoid()
+    local medoid2 = pattern.new({{1,0,1},
+                                 {0,1,0},
+                                 {1,0,1}}):medoid()
+    local medoid3 = pattern.new({{0,1,0},
+                                 {0,1,0},
+                                 {0,1,0}}):medoid()
+    local medoid4 = pattern.new({{0,0,0},
+                                 {1,1,1},
+                                 {0,0,0}}):medoid()
+    local medoid5 = pattern.new({{0,0,0},
+                                 {0,1,0},
+                                 {0,0,0}}):medoid()
+    -- This should not be the same as the others
+    local medoid6 = pattern.new({{0,1,0},
+                                 {1,0,1},
+                                 {0,1,0}}):medoid()
+    lu.assertEquals(medoid1, medoid2)
+    lu.assertEquals(medoid2, medoid3)
+    lu.assertEquals(medoid3, medoid4)
+    lu.assertEquals(medoid4, medoid5)
+    lu.assertNotEquals(medoid5, medoid6)
 end
 
 function testPattern:testSurface()